added wrapper for variant context builder
