+++ VariantCaller/src/main/java/com/epam/bioinf/variantcaller/exceptions/caller/NoGenotypesException.java
@@ -1,0 +1,17 @@
+package com.epam.bioinf.variantcaller.exceptions.caller;
+
+import com.epam.bioinf.variantcaller.exceptions.HandlerException;
+
+/**
+ * Thrown when something goes wrong while processing region data.
+ */
+public class NoGenotypesException extends HandlerException {
+  public NoGenotypesException(String s) {
+    super(s);
+  }
+
+  public NoGenotypesException(String s, Throwable throwable) {
+    super(s, throwable);
+  }
+}
+
